buy SkillUp in ProgressSystem

+++ Assets/Scripts/ProgressSystem/MoneySystem.cs
@@ -3,9 +3,9 @@ using System;
 
 public class MoneySystem
 {
-    public int Money { get; private set; }
+    public float Money { get; private set; }
 
-    public event Action<int> MoneyChange;
+    public event Action<float> MoneyChange;
 
     public MoneySystem(EnemySpawnManager enemySpawnManager) => 
         enemySpawnManager.OnSpawned += AddEnemy;
@@ -13,13 +13,13 @@ public class MoneySystem
     private void AddEnemy(EnemyDeath enemy) =>
         enemy.OnDeadMonty += AddMoney;
 
-    public void AddMoney(int value)
+    public void AddMoney(float value)
     {
         Money += value;
         MoneyChange?.Invoke(Money);
     }
 
-    public void SpendMoney(int value)
+    public void SpendMoney(float value)
     {
         Money -= value;
         MoneyChange?.Invoke(Money);
